Password Verification Correctly Implemented
AppleScript wants to only run shell commands as root using its with administrator priveleges. Which just uses granted permissions to the app to authenticate. This is super unhelpful when you are trying to verify the entered password.

This adds in a simple shell script file that runs a command that requries the current user password, just to check if the entered password is correct.